backend/pyproject.toml
[project]
name = "voice-ide-backend"
version = "0.1.0"
description = "Local backend for Voice IDE (files/tools/agent/voice)"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "python-dotenv>=1.0",
  "pydantic>=2.6",
]

[tool.uvicorn]
factory = false
